+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 - Added ``fields`` parameter. `#34`_
 - Added retrying support on rate limit exceeding. `#43`_
 - Better exceptions representation. `#52`_
+- Added type checks for calls parameters. `#51`_
 
 0.4.4 (2016-09-09)
 ------------------
@@ -105,6 +106,7 @@ Changelog
 
 
 .. _#52: https://github.com/Stranger6667/pyoffers/issues/52
+.. _#51: https://github.com/Stranger6667/pyoffers/issues/51
 .. _#45: https://github.com/Stranger6667/pyoffers/issues/45
 .. _#43: https://github.com/Stranger6667/pyoffers/issues/43
 .. _#42: https://github.com/Stranger6667/pyoffers/issues/42